--- app/components/ProductList.tsx/ProductList.test.ts
@@ -1,3 +1,0 @@
-test('test', () => {
-  expect(true).toBe(true);
-});
